# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Título de tu API",
        "description": "Descripción de tu API",
        "contact": {
            "name": "Nombre del Contacto",
            "email": "contacto@ejemplo.com"
        },
        "license": {
            "name": "Licencia",
            "url": "http://www.ejemplo.com/licencia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/decks/{playerId}/{deckId}": {
            "get": {
                "tags": [
                    "Mazos"
                ],
                "summary": "Recuperar un mazo por ID de jugador y ID de mazo",
                "operationId": "a19410095b687dbb7e9c7ab5495fe136",
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mazo obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deck fetched successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mazo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deck not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocurrió un error al obtener el mazo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred while fetching the deck"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Mazos"
                ],
                "summary": "Actualiza un mazo existente para un jugador específico",
                "operationId": "2c74d32bbbb92fc740d788039b7cb426",
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "deckId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_card_1": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_card_2": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "id_card_3": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "id_card_4": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "id_card_5": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "id_card_6": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "id_card_7": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "id_card_8": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "id_deck_player": {
                                        "description": "Nombre del mazo",
                                        "type": "string",
                                        "example": "Deck1"
                                    },
                                    "user_id": {
                                        "description": "ID del usuario propietario del mazo",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mazo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mazo actualizado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación o al actualizar el mazo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "id_card_1": "Este campo es obligatorio"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El mazo no se encontró",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El mazo con el ID especificado no se encontró."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/decks/{playerId}": {
            "post": {
                "tags": [
                    "Mazos"
                ],
                "summary": "Crea un nuevo mazo para un jugador específico",
                "operationId": "8d7082790da6c52f1e1afd4084ddc275",
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_card_1": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_card_2": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "id_card_3": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "id_card_4": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "id_card_5": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "id_card_6": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "id_card_7": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "id_card_8": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "id_deck_player": {
                                        "type": "string",
                                        "example": "Deck1"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mazo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mazo creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación o al crear el mazo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "id_card_1": "Este campo es obligatorio"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}